fix(provider/telegram): pre-validate group children + share identity mappers
Address ninth CodeRabbit review on PR #16.

- messages.ts (sendGroupContent): scan all group items for unsupported
  child types (group, reaction, custom, poll_option, effect) BEFORE
  posting any of them. The previous loop dispatched each child as it
  iterated, so a heterogeneous group like [photo, reaction, contact]
  would push the photo to Telegram first and then throw on `reaction`,
  leaving an orphan media message in the chat with no group context
  and no atomic undo. Pre-validation reuses the same UnsupportedError
  shapes (and detail strings) that dispatchSend would have thrown,
  preserving the user-facing contract — only the *timing* of the
  throw moves earlier.

- identity.ts (new): hoist chatToSpace, userToSender, chatToSender
  into a fresh leaf module so both inbound (events/inbound.ts) and
  outbound (messages.ts) paths share the *exact* same mapping logic.
  Previously messages.ts re-implemented buildOutboundSender and
  buildChatSender locally to avoid a circular import with
  events/inbound.ts (which imports toGroupItems from messages.ts).
  Centralizing eliminates the recurring drift hazard CodeRabbit has
  flagged twice — when chatToSender gained a sender_chat fallback,
  buildChatSender had to be added in parallel; with one source the
  next change can't desync.

- events/inbound.ts, events/polls.ts, events/reactions.ts: import
  identity helpers from ../identity directly (no re-export hop
  through inbound.ts; biome's noExportedImports disallows that).

- messages.ts: replace local buildOutboundSender/buildChatSender and
  the inline space builder with calls to the shared mappers.

Verified: tsc --noEmit clean, bun x ultracite check clean across all
90 files.
